backend/requirements.txt
# Flask and extensions
Flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4

# Database
SQLAlchemy==2.0.9
alembic==1.10.3

# Utilities
python-dotenv==1.0.0
Werkzeug==2.2.3
requests==2.28.2
pyjwt==2.6.0

# AI and RAG
langchain==0.0.267
openai==1.3.0
faiss-cpu==1.7.4
tiktoken==0.4.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2

# Testing
pytest==7.3.1
pytest-flask==1.2.0 